move github assignment grader
